split perf into separate modules
clang (and maybe g++) was optimising far too well.
Splitting functions into their own module makes it easier to see the
resultant code in the .s.

as for the main perf module, used timing functions from an array of,
which prevented clang from optimising through the fn ptr call.
so now the ref timings are similar in clang/g++.