optional: improve efficiency for large T
optional wan't that efficient for large types,.
The Some/None classes caused most of the issue and were extra code.

Removed the Some/None classes, reducing code and allowing more
efficient inplace construct.

Also removed the this-> as this is not idiomatic c++

As no longer using a None class, now don't need the
maybe-uninitialised silencing..