Optional/Result: Allow use of references as stored type.
The current implementation of Result and Optional does not allow
references to be stored.
References are needed to allow the compiler to reduce the amount of
copying/moving that needs to be done when iterating over a collection.

This commit adjusts optional and result to take reference types as the
stored type.