vector: make try_reserve an asb behaviour
reserve was using a relative behaviour, i.e. new reservation was
additional to the current, which makes it hard to reduce.

The new one uses absolute behaviour, so now can reduce.
Also, new one uses quanta based allocation so will over-alloc.

In doing this found segfaults in the reallocator and fixed them.