result+optional: Handle inval case better.
The inval state occurs when value is moved away/consumed or not yet set.
but in these cases it is inval and must not be used.
So panic if being used in these cases.
(even for predicates.. if inval when checking for somes then something
is broken).

Also, find out that std::function adds bloat to the emitted code, so
remove it.
Then find that template function para specifiers don't work any more and
need a 'template'  prefix.. which is butt ugly.