all: Strip out noexcept clauses/
noexcept is a clause that says the function does not throw.
noexcept also says that it blocks any exception bubbling up
from called functions, so the compiler will emit exception
detection and blocking code if exceptions are enabled.

Which is not what's wanted.

What is wanted is functions must not raise their own exceptions
but can propagate exceptions from called functions.