build: Test with separate build.
The test phase is using the artifacts from the debug build.
But testing has different flags active than debug.
Which means it's inconsistent.

So, remove the use of debug build and build tests with their
own settings.

Which allows #if defined(TEST) usage inn the main code.
Arguably not wanted but needed.